Created SwitchBar UI. Last commit before a long pause.
Starting from this commit there will be a long pause. This project will be in pause until my last academic year has completed (about June 2025), unless some exceptional times during this period. For the next time: continue the switchbar adding a controller that, based on the switchbar piece touched, the main camera will translate her position on the correct screen side. Follow this, create the "collection section" with the joker's button, and to create in some way a grid of collectables, showing only the joker's found in the packs, the "padlock" image otherwise.
And please, don't quit me!